Phase F2/F3: fused int immediates and Copy instructions
Emit SubImm/LeImm/etc. for x op literal, make Instruction Copy with
u32 line, and shave Fib(30) to ~273ms (~2.13x baseline).

+++ crates/mailang-bytecode/src/format.rs
@@ -461,7 +461,7 @@ pub fn decode(data: &[u8]) -> Result<Bytecode, BytecodeFormatError> {
             } else {
                 None
             };
-            let line = r.u32()? as usize;
+            let line = r.u32()?;
             instructions.push(Instruction {
                 opcode,
                 operand,
